Grandes Updates!
- Mudança na aparencia do sistema
- Tela de cadastro de professor feita
- Controllers para cadastrar o professor feitos
- Falta o controller para consulta de professor e para remover professor (usar HashTable nisso)

+++ src/controller/IInscricao.java
@@ -1,7 +1,7 @@
 package controller;
 
 public interface IInscricao {
-	public void novaInscricao(int cpfProfessor, int codigoDisciplina, int codigoProcesso);
-	public void salvarInscricao(int cpfProfessor, int codigoDisciplina, int codigoProcesso);
-
+	public void novaInscricao(long cpfProfessor, int codigoDisciplina, int codigoProcesso);
+	public void salvarInscricao(long cpfProfessor, int codigoDisciplina, int codigoProcesso);
+	
 }
